backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-search-views==0.3.7
djangorestframework==3.12.2
libsass==0.20.1
mysqlclient==2.0.3
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
